environment profiler is now outside of setup teardown scripts, as well as vm details helper.
also, we add reservation id to filename
