+++ environment_scripts/env_setup/setup_script.py
@@ -1,10 +1,11 @@
 from multiprocessing.pool import ThreadPool
-import os
-import cProfile
-import pstats
 import cloudshell.api.cloudshell_scripts_helpers as helpers
 from cloudshell.api.cloudshell_api import *
 from cloudshell.core.logger import qs_logger
+
+from environment_scripts.helpers.vm_details_helper import get_vm_custom_param
+from environment_scripts.profiler.env_profiler import profileit
+
 
 class EnvironmentSetup:
     def __init__(self):
@@ -149,34 +150,3 @@ class EnvironmentSetup:
             return False
 
         return True
-
-
-def get_vm_custom_param(vm_custom_params, param_name):
-    """
-    :param list[VmCustomParam] vm_custom_params:
-    :param param_name:
-    :return:
-    """
-    for param in vm_custom_params:
-        if param.Name == param_name:
-            return param
-    return None
-
-
-### http://stackoverflow.com/questions/5375624/a-decorator-that-profiles-a-method-call-and-logs-the-profiling-result ###
-def profileit(scriptName):
-    def inner(func):
-        import cloudshell.api.cloudshell_scripts_helpers as helpers
-        profiling = helpers.get_global_inputs().get('quali_profiling')
-        environment_name = helpers.get_reservation_context_details().environment_name
-        def wrapper(*args, **kwargs):
-            if not profiling:
-                return func(*args, **kwargs)
-            prof = cProfile.Profile()
-            retval = prof.runcall(func, *args, **kwargs)
-            s = open(os.path.join(profiling, scriptName + "_" + environment_name + ".text"), 'w')
-            stats = pstats.Stats(prof, stream=s)
-            stats.strip_dirs().sort_stats('cumtime').print_stats()
-            return retval
-        return wrapper
-    return inner
